뷰 분리
