+++ backend/finance_manager/views/asset_views.py
@@ -1,215 +1,24 @@
-# from django.shortcuts import render, redirect, get_object_or_404, get_list_or_404
-# from django.http import HttpResponse, JsonResponse
-# from django.core import serializers
-# from ..forms import AssetForm
-# from django.contrib.auth.decorators import login_required
+# from django.http import Http404
+# from rest_framework.views import APIView
+# from rest_framework.response import Response
+# from rest_framework import status
 
-from django.views.decorators.csrf import csrf_exempt
-from rest_framework import permissions, viewsets, serializers
-from rest_framework.response import Response
-from rest_framework.decorators import api_view
+from . import LargeResultsSetPagination, BaseViewSet
 
-from finance_manager.models.asset import Asset, USStock, KOStock
-from finance_manager.serializers.asset import (
-    USStockSerializer,
-    KOStockSerializer,
-    AssetSerializer,
-    ExpectAssetSerializer,
-)
-from . import LargeResultsSetPagination
+from finance_manager.models.portfolio import Portfolio
+from finance_manager.models.asset import Stock, Realty
+
+from finance_manager.serializers.portfolio_srzs import PortfolioSerializer
+from finance_manager.serializers.asset_srzs import StockSerializer, RealtySerializer
 
 
-class AssetScheduler:
-    # TODO: 값 검증
-    def __init__(
-        self, present_value, monthly_input, interest, future_value=None, year=None
-    ):
-        self.pv = present_value
-        self.fv = future_value
-        self.a = monthly_input
-        self.r = interest * 0.01 + 1
-        self.n = year
-        self.data = {}
+class StockViewSet(BaseViewSet):
 
-    def calculate(self):
-        """
-        빈 파라미터를 채우는 연산 적용
-        """
-        if self.fv is None:
-            self._calculate_fv()
-
-        elif self.n is None:
-            self._calculate_y()
-
-    def _calculate_fv(self):
-        """
-        투입금, 이자율, 기간 -> 미래 가치 계산
-        """
-        pv = self.pv
-        a = self.a
-        r = self.r
-        n = self.n
-        fvs = {}
-
-        for i in range(1, n + 1):
-            fvs[i] = {}
-
-            # 기대 자산
-            fvs[i]["future_value"] = (
-                round(pv * r + a)
-                if i == 1  # 첫해 투입분 이자 x
-                else round((pv * (r ** i)) + (a * (r ** i - 1) / (r - 1)))
-            )
-
-            # 누적 추가 투자액
-            fvs[i]["cum_input"] = a * i
-
-            # 누적 이자 = 기대 자산 - (현재 가치 + 누적 추가 투자액)
-            fvs[i]["cum_interest"] = fvs[i]["future_value"] - (pv + fvs[i]["cum_input"])
-
-        self.data["future_values"] = fvs
-        # 마지막 값 바로 찾을 수 있도록 추가
-        self.data["future_value"] = self.data["future_values"][n]["future_value"]
-
-    def _calculate_y(self):
-        """
-        미래가치, 투입금, 이자율 -> 기간 계산
-        """
-        # import numpy as np
-
-        pv = self.pv
-        fv = self.fv
-        a = self.a
-        r = self.r
-
-        i, temp_fv = 1, 0
-        fvs = {}
-        # direct
-        # n = np.log(r) / np.log(1 - (1 - r) * fv / a)
-
-        # TODO: 반복제거
-        while fv > temp_fv:
-            fvs[i] = {}
-
-            # 기대 자산
-            temp_fv = (
-                round(pv * r + a)
-                if i == 1  # 첫해 투입분 이자 x
-                else round((pv * (r ** i)) + (a * (r ** i - 1) / (r - 1)))
-            )
-            fvs[i]["future_value"] = temp_fv
-            fvs[i]["cum_input"] = a * i
-            fvs[i]["cum_interest"] = fvs[i]["future_value"] - (pv + fvs[i]["cum_input"])
-            i += 1
-
-        self.data["future_values"] = fvs
-        self.data["period"] = i - 1
+    queryset = Stock.objects.all()
+    serializer_class = StockSerializer
 
 
-@api_view(["POST", "GET"])
-@csrf_exempt
-def get_expect_asset(request):
-    if request.method == "GET":
-        serializer = ExpectAssetSerializer()
-        return Response(serializer.data)
-    else:
+class RealtyViewSet(BaseViewSet):
 
-        data = request.data
-
-        asset_scheduler = AssetScheduler(**data)
-        asset_scheduler.calculate()
-
-        data["expect"] = asset_scheduler.data
-
-        return Response(data=data)
-
-
-class BaseViewSet(viewsets.ModelViewSet):
-
-    permission_classes = [permissions.IsAuthenticatedOrReadOnly]
-    pagination_class = LargeResultsSetPagination
-
-    def perform_create(self, serializer):
-        serializer.save(user=self.request.user)
-
-
-class USStockViewSet(BaseViewSet):
-
-    queryset = USStock.objects.all()
-    serializer_class = USStockSerializer
-
-
-class KOStockViewSet(BaseViewSet):
-
-    queryset = KOStock.objects.all()
-    serializer_class = KOStockSerializer
-
-
-class AssetViewSet(BaseViewSet):
-
-    queryset = Asset.objects.all()
-    serializer_class = AssetSerializer
-
-
-# @login_required(login_url="common:login")
-# def asset_create(request):
-#     """
-#     자산 등록
-#     """
-#     if request.method == "POST":
-#         form = AssetForm(request.POST)
-#         if form.is_valid():
-#             asset = form.save(commit=False)
-#             asset.user = request.user
-#             asset.save()
-#             return redirect("/")
-#     else:
-#         form = AssetForm()
-
-#     context = {"form": form}
-
-#     return render(request, "finance_manager/asset_form.html", context)
-
-
-# @login_required(login_url="common:login")
-# def asset_read(request):
-#     """
-#     자산 출력
-#     """
-#     # TODO: generic - table view
-#     asset = Asset.objects.filter(user=request.user)
-#     asset = get_list_or_404(asset)
-#     # serialized_obj = serializers.serialize(
-#     #     "json",
-#     #     [asset],
-#     # )
-#     return HttpResponse(asset)
-
-
-# @login_required(login_url="common:login")
-# def asset_modify(request, asset_id):
-#     """
-#     자산 수정
-#     """
-#     asset = get_object_or_404(Asset, pk=asset_id)
-
-#     if request.method == "POST":
-#         form = AssetForm(request.POST, instance=asset)
-#         if form.is_valid():
-#             asset = form.save(commit=False)
-#             asset.save()
-#             return redirect("/")
-#     else:
-#         form = AssetForm(instance=asset)
-#     context = {"form": form}
-#     return render(request, "finance_manager/asset_form.html", context)
-
-
-# @login_required(login_url="common:login")
-# def asset_delete(request, asset_id):
-#     """
-#     자산 삭제
-#     """
-#     asset = get_object_or_404(Asset, pk=asset_id)
-#     asset.delete()
-#     return redirect("/")
+    queryset = Realty.objects.all()
+    serializer_class = RealtySerializer
